[BE/fix] 파일 읽어오는 걸로 변경 / decimal->float

+++ backend/src/main/java/ssafy/haruman/domain/deposit/dto/request/DepositCreateRequestDto.java
@@ -4,8 +4,6 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class DepositCreateRequestDto {
@@ -13,6 +11,6 @@ public class DepositCreateRequestDto {
     private String bank;
     private String name;
     private String description;
-    private BigDecimal interestRate;
+    private float interestRate;
 
 }
