+++ backend/src/main/java/ssafy/haruman/domain/deposit/dto/response/DepositSimpleResponseDto.java
@@ -3,8 +3,6 @@ package ssafy.haruman.domain.deposit.dto.response;
 import lombok.*;
 import ssafy.haruman.domain.deposit.entity.Deposit;
 
-import java.math.BigDecimal;
-
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
@@ -13,7 +11,7 @@ public class DepositSimpleResponseDto {
 
     private String bank;
     private String name;
-    private BigDecimal interestRate;
+    private float interestRate;
 
     public static DepositSimpleResponseDto from(Deposit deposit) {
         return DepositSimpleResponseDto.builder()
